requirements.txt
Django==5.2.2
djangorestframework==3.15.2
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
Pillow==10.4.0
django-cors-headers==4.3.1
reportlab==4.0.9
openpyxl==3.1.2
python-docx==1.1.0
django-filter==23.5
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-bootstrap5==23.3
djangorestframework-simplejwt==5.3.0
python-dateutil==2.9.0
pytz==2024.1
django-import-export==4.0.3
django-cleanup==8.0.2